+++ src/app/cpp/PageDefaults.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "core/model/PageStyle.hpp"
+
+namespace phvikapen::app::defaults {
+
+[[nodiscard]] core::PageStyle pageStyle();
+
+void setPageStyle(const core::PageStyle& style);
+
+}